+++ civgame/__init__.py
@@ -7,7 +7,7 @@ from .constants import (
     GAME_CONFIG,
     CITY_NAMES,
 )
-from .data import TECHNOLOGIES, UNIT_TYPES, BUILDINGS, CIVILIZATIONS, IMPROVEMENTS
+from .data import TECHNOLOGIES, UNIT_TYPES, BUILDINGS, CIVILIZATIONS, IMPROVEMENTS, RESOURCES
 from .hex import hex_neighbors, hex_distance, offset_to_cube
 from .mapgen import generate_map, generate_earth_map
 from .state import GameState
@@ -16,7 +16,7 @@ __all__ = [
     "Terrain",
     "TERRAIN_YIELDS", "TERRAIN_MOVE_COST", "TERRAIN_DEFENSE",
     "GAME_CONFIG", "CITY_NAMES",
-    "TECHNOLOGIES", "UNIT_TYPES", "BUILDINGS", "CIVILIZATIONS", "IMPROVEMENTS",
+    "TECHNOLOGIES", "UNIT_TYPES", "BUILDINGS", "CIVILIZATIONS", "IMPROVEMENTS", "RESOURCES",
     "hex_neighbors", "hex_distance", "offset_to_cube",
     "generate_map", "generate_earth_map",
     "GameState",